requirements.txt
gradio>=4.44.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
sentence-transformers>=2.7.0
pypdf>=4.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
